+++ src/losses.jl
@@ -13,10 +13,10 @@ function likelihood_loss(μ::Matrix{<:Real}, σ::Matrix{<:Real}, pi::Matrix{<:Re
     return likelihood_loss(Float32.(μ), Float32.(σ), Float32.(pi), Float32.(y))
 end
 
-function likelihood_loss(μ::Matrix{Float32}, σ::Matrix{Float32}, pi::Matrix{Float32}, y::Matrix{Float32})
-	@pipe pi .* (1.0f0 ./ ((sqrt(Float32(2π)) .* σ)).*exp.(-0.5f0((y .- μ) ./ σ).^2.0f0)) |>
+function likelihood_loss(μ::Matrix{Float64}, σ::Matrix{Float64}, pi::Matrix{Float64}, y::Matrix{Float64})
+	@pipe pi .* (1.0 ./ ((sqrt(2π) .* σ)).*exp.(-0.5((y .- μ) ./ σ).^2.0)) |>
 	sum(_, dims=1) |>
-	clamp.(_, eps(Float32), 10000.0f0) |>
+	clamp.(_, eps(Float64), 10000.0) |>
 	-log.(_) |>
 	mean(_)
 end
